Starts the implementation of the dielectric material.

+++ rain/include/scene/components/shape/materials/metal_material.h
@@ -20,7 +20,7 @@ class MetalMaterial : public LambertianMaterial {
          * @param albedo Albedo
          * @param fuzz
          */
-        MetalMaterial(Vec3 albedo, int fuzz) :
+        MetalMaterial(Vec3 albedo, float fuzz) :
             LambertianMaterial(albedo), fuzz(fuzz) { /* empty */ }
 
         /*!
